# Conflicts:
#	solr/api/gradle.lockfile
#	solr/benchmark/gradle.lockfile
#	solr/core/gradle.lockfile
#	solr/cross-dc-manager/gradle.lockfile
#	solr/modules/analysis-extras/gradle.lockfile
#	solr/modules/clustering/gradle.lockfile
#	solr/modules/cross-dc/gradle.lockfile
#	solr/modules/cuvs/gradle.lockfile
#	solr/modules/extraction/gradle.lockfile
#	solr/modules/gcs-repository/gradle.lockfile
#	solr/modules/jwt-auth/gradle.lockfile
#	solr/modules/langid/gradle.lockfile
#	solr/modules/language-models/gradle.lockfile
#	solr/modules/ltr/gradle.lockfile
#	solr/modules/opentelemetry/gradle.lockfile
#	solr/modules/s3-repository/gradle.lockfile
#	solr/modules/scripting/gradle.lockfile
#	solr/modules/sql/gradle.lockfile
#	solr/solr-ref-guide/gradle.lockfile
#	solr/solrj-jetty/gradle.lockfile
#	solr/solrj-streaming/gradle.lockfile
#	solr/solrj-zookeeper/gradle.lockfile
#	solr/solrj/gradle.lockfile
#	solr/test-framework/gradle.lockfile

+++ solr/core/src/test/org/apache/solr/handler/admin/api/GetMetricsTest.java
@@ -87,8 +87,14 @@ public class GetMetricsTest extends SolrTestCaseJ4 {
 
   @AfterClass
   public static void afterClass() throws Exception {
-    jettyHttpClient.destroy();
-    cluster.shutdown();
+    if (jettyHttpClient != null) {
+      jettyHttpClient.destroy();
+      jettyHttpClient = null;
+    }
+    if (cluster != null) {
+      cluster.shutdown();
+      cluster = null;
+    }
   }
 
   @Before
